+++ crates/swc_core/src/quote.rs
@@ -1,6 +1,3 @@
-use self::private::Sealed;
-use super::{ast::*, atoms::JsWord};
-
 /// # Supported output types
 ///
 ///  - `Expr`
@@ -88,34 +85,15 @@ macro_rules! quote_expr {
 
     ($src:tt, $($tt2:tt)*) => {{
         $crate::quote!($src as Box<Expr>, $($tt2)*)
+
     }};
 }
 
-/// Not a public API.
-///
-/// Implemented for types which will be implicitly cloned when used as a
-/// variable in [crate::quote] macro calls.
-pub trait ImplicitClone: Clone + Sealed {
+/// Noop
+pub trait ImplicitClone: Clone {
     fn clone_quote_var(&self) -> Self {
         self.clone()
     }
 }
 
-macro_rules! impl_for {
-    ($T:ty) => {
-        impl ImplicitClone for $T {}
-        impl Sealed for $T {}
-    };
-}
-
-impl_for!(Id);
-impl_for!(Ident);
-impl_for!(JsWord);
-
-impl_for!(Str);
-impl_for!(Number);
-impl_for!(Bool);
-
-mod private {
-    pub trait Sealed {}
-}
+impl<T: Clone> ImplicitClone for T {}